Modulate the Button, Card, and Input Components
