+++ src/components/ui/Card.jsx
@@ -1,0 +1,4 @@
+export function Card({ children, className }) {
+    return <div className={`p-4 border rounded bg-gray-800 ${className}`}>{children}</div>;
+  }
+
